nxp-flexspi: add support for channel/CS addressing
Rework driver to use 'reg' property for specifying
channel and CS number. Values 0 and 1 correspond
to channel A (CS0 and CS1 respectively), 2 and 3 to channel B.

Handling of 'fsl,has-second-chip' property is removed from
driver as it doesn't make much sense in this new scheme
where presence or absence of devices on CS1 can be expressed
directly using 'reg'.

Signed-off-by: Edgar Cherkasov <echerkasov@dev.rtsoft.ru>
Signed-off-by: Michael Walle <michael.walle@kontron.com>